renamed auth and cart files
