+++ projects/core/src/auth/store/selectors/auth-group.selectors.ts
